add custom exception handling for authorization and forbidden actions

+++ tdd-bdd/src/main/java/com/pandorian/tdd_bdd/controllers/NoteController.java
@@ -4,6 +4,7 @@ import com.pandorian.tdd_bdd.annotations.CurrentUser;
 import com.pandorian.tdd_bdd.entity.Note;
 import com.pandorian.tdd_bdd.entity.User;
 import com.pandorian.tdd_bdd.exceptions.RequiredArgumentMissingException;
+import com.pandorian.tdd_bdd.exceptions.UserNotAuthorized;
 import com.pandorian.tdd_bdd.service.NoteService;
 import org.hibernate.Hibernate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/notes")
@@ -18,6 +20,12 @@ public class NoteController {
 
     @Autowired
     private NoteService noteService;
+
+    void checkAuthority(boolean authorized) {
+        if (!authorized) {
+            throw new UserNotAuthorized();
+        }
+    }
 
     @AuthRequired
     @GetMapping
@@ -51,8 +59,9 @@ public class NoteController {
         return ResponseEntity.status(201).body(note);
     }
 
+    @AuthRequired
     @PutMapping("/{id}")
-    public ResponseEntity<?> modifyNote(@PathVariable int id, @RequestBody Map<String, String> payload) {
+    public ResponseEntity<?> modifyNote(@PathVariable int id, @RequestBody Map<String, String> payload, @CurrentUser User user) {
         String title = payload.get("title");
         String content = payload.get("content");
 
@@ -65,6 +74,8 @@ public class NoteController {
 
         Note note = noteService.getNoteById(id);
 
+        checkAuthority(Objects.equals(user.getId(), note.getOwner().getId()));
+
         note.setTitle(title);
 
         note.setContent(content);
@@ -74,8 +85,14 @@ public class NoteController {
         return ResponseEntity.ok(note);
     }
 
+    @AuthRequired
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteNote(@PathVariable int id) {
+    public ResponseEntity<?> deleteNote(@PathVariable int id, @CurrentUser User user) {
+
+        Note note = noteService.getNoteById(id);
+
+        checkAuthority(Objects.equals(user.getId(), note.getOwner().getId()));
+
         noteService.deleteNoteById(id);
         return ResponseEntity.noContent().build();
     }
